requirements.txt
flask==3.0.3
flask-cors==4.0.1
boto3==1.34.0
python-pptx==0.6.23
werkzeug==3.0.3
gunicorn==22.0.0
